LayoutTests/accessibility/aria-labelledby-on-input.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test-pre.js"></script>
<script src="../resources/accessibility-helper.js"></script>
</head>
<body>
<div id="content">
    <span id="message">This computer will self-destruct in</span>
    <input id="time" type="text" value="10" aria-labelledby="message time unit"/>
    <span id="unit"> minutes.</span>
</div>
<div id="console"></div>
<script>
    description("This verifies the accessible name of an input with multiple aria-labelledby elements.");
    if (window.accessibilityController) {
        var labeledItem = document.getElementById("time").focus();
        var axInput = accessibilityController.focusedElement;
        shouldBeEqualToString("platformValueForW3CName(axInput)", "This computer will self-destruct in 10 minutes.");

        document.getElementById("content").style.visibility = "hidden";
    }
</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
