Safari-611.1.10/LayoutTests/accessibility/aria-labelledby-on-input-expected.txt

This verifies the accessible name of an input with multiple aria-labelledby elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS platformValueForW3CName(axInput) is "This computer will self-destruct in 10 minutes."
PASS successfullyParsed is true

TEST COMPLETE
